+++ R/Exploratory-methods.R
@@ -87,6 +87,8 @@ setMethod(
 #' @param suppress a numeric value indicating which (possibly rotated) factor
 #'   loadings should be suppressed. Typical values are around .3 in most
 #'   statistical software. Default is 0 for no suppression
+#' @param printCI print a confidence interval for standardized loadings
+#'   (e.g., \code{printCI = .95} gives a 95\% confidence interval)
 #' @param digits number of significant digits to be rounded
 #' @param verbose logical; allow information to be printed to the console?
 #' @param ... additional arguments to be passed
@@ -103,12 +105,16 @@ setMethod(
 #' x <- mirt(Science, 2)
 #' summary(x)
 #' summary(x, rotate = 'varimax')
+#' 
+#' #print confidence interval (requires computed information matrix)
+#' x2 <- mirt(Science, 1, SE=TRUE)
+#' summary(x2, printCI=.95)
 #' }
 setMethod(
     f = "summary",
     signature = 'ExploratoryClass',
     definition = function(object, rotate = '', Target = NULL, suppress = 0, digits = 3,
-                          verbose = TRUE, ...){
+                          printCI = FALSE, verbose = TRUE, ...){
         nfact <- ncol(object@F)
         if (rotate == 'none' || nfact == 1) {
             F <- object@F