+++ R/utils.R
@@ -232,13 +232,40 @@ test_info <- function(pars, Theta, Alist, K){
     info
 }
 
-Lambdas <- function(pars){
-    lambdas <- list()
-    J <- ifelse(is(pars[[length(pars)]], 'GroupPars'), length(pars)-1L, length(pars))
-    for(i in 1L:J)
-        lambdas[[i]] <- ExtractLambdas(pars[[i]])
-    lambdas <- do.call(rbind,lambdas)
-    lambdas
+Lambdas <- function(pars, Names, explor = FALSE, alpha = .05){
+    J <- length(pars) - 1L
+    lambdas <- lowerlambdas <- upperlambdas <- 
+        matrix(NA, J, length(ExtractLambdas(pars[[1L]])))
+    z <- qnorm(1 - alpha/2)
+    rownames(lambdas) <- rownames(upperlambdas) <- rownames(lowerlambdas) <- Names
+    for(i in 1L:J){
+        tmp <- pars[[i]]
+        lambdas[i,] <- ExtractLambdas(tmp)/1.702
+        tmp@par <- pars[[i]]@par - z * pars[[i]]@SEpar
+        lowerlambdas[i,] <- ExtractLambdas(tmp)/1.702
+        tmp@par <- pars[[i]]@par + z * pars[[i]]@SEpar
+        upperlambdas[i,] <- ExtractLambdas(tmp)/1.702
+    }
+    if(ncol(lambdas) > 1L){
+        norm <- sqrt(1 + rowSums(lambdas^2))
+    } else {
+        norm <- as.matrix(sqrt(1 + lambdas[ ,1L]^2))
+    }
+    F <- as.matrix(lambdas/norm)
+    if(!explor){
+        if(ncol(lambdas) > 1L){
+            norml <- sqrt(1 + rowSums(lowerlambdas^2))
+            normh <- sqrt(1 + rowSums(upperlambdas^2))
+        } else {
+            norml <- as.matrix(sqrt(1 + lowerlambdas[ ,1L]^2))
+            normh <- as.matrix(sqrt(1 + upperlambdas[ ,1L]^2))
+        }
+        ret <- list(F=F, lower=as.matrix(lowerlambdas/norml), 
+                    upper=as.matrix(upperlambdas/normh))
+    } else {
+        ret <- list(F=F, lower=list(), upper=list())
+    }
+    ret
 }
 
 #change long pars for groups into mean in sigma